requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-ollama
ollama
python-dotenv